add gemspec
support installation via ruby gems for environments like vagrant or
docker where the emacs instance might not be running on the same OS as
the inf-ruby console.

This allows to install the gem via ruby gems and require it this way, so
robe-mode can be used without access to the full host os Filesystem.